[refactor] Move `TestStore+Extensions` to single file
